Add test for simulation of no kconfigs in system
https://app.codecov.io/gh/a13xp0p0v/kernel-hardening-checker/blob/master/kernel_hardening_checker%2F__init__.py?flags%5B0%5D=functional_test#L68

+++ .github/workflows/functional_test.sh
@@ -196,6 +196,23 @@ coverage run -a --branch bin/kernel-hardening-checker -g X86_64 -m show_ok && ex
 echo ">>>>> no kconfig file <<<<<"
 coverage run -a --branch bin/kernel-hardening-checker -c ./nosuchfile && exit 1
 
+echo ">>>>> no kconfig file for autodetection <<<<<"
+FILE3="/proc/config.gz"
+FILE4="/boot/config-$(uname -r)"
+if [ -f "$FILE3" ]; then
+    echo "$FILE3 exists, cannot simulate the absence of a kconfig"
+    exit 1
+fi
+if [ -f "$FILE4" ]; then
+    echo "$FILE4 exists, hiding it temporarily"
+    sudo mv "$FILE4" "$FILE4.bak"
+    ret=0; coverage run -a --branch bin/kernel-hardening-checker -a || ret=$? # check the test result after restoring /boot/config-*
+    sudo mv "$FILE4.bak" "$FILE4"
+    [ $ret -eq 0 ] && exit 1
+else
+    coverage run -a --branch bin/kernel-hardening-checker -a && exit 1
+fi
+
 echo ">>>>> no kernel version <<<<<"
 sed '3d' test.config > error.config
 coverage run -a --branch bin/kernel-hardening-checker -c error.config && exit 1
